DL-PokemonClassification/api/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
tensorflow-macos>=2.16.0
tensorflow-metal>=1.1.0
pillow>=10.0.0
numpy>=1.26.0
